+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/integrate/simple/SimpleElasticJobTest.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,14 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.integrate.std.dataflow;
+package org.apache.shardingsphere.elasticjob.lite.integrate.simple;
 
 import org.apache.shardingsphere.elasticjob.lite.api.ElasticJob;
 import org.apache.shardingsphere.elasticjob.lite.api.JobType;
 import org.apache.shardingsphere.elasticjob.lite.config.JobConfiguration;
-import org.apache.shardingsphere.elasticjob.lite.executor.type.impl.DataflowJobExecutor;
 import org.apache.shardingsphere.elasticjob.lite.integrate.EnabledJobIntegrateTest;
-import org.apache.shardingsphere.elasticjob.lite.integrate.fixture.dataflow.OneOffDataflowElasticJob;
+import org.apache.shardingsphere.elasticjob.lite.integrate.fixture.simple.FooSimpleElasticJob;
 import org.apache.shardingsphere.elasticjob.lite.util.concurrent.BlockUtils;
 import org.junit.After;
 import org.junit.Before;
@@ -30,27 +29,26 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
 
-public final class OneOffDataflowElasticJobTest extends EnabledJobIntegrateTest {
+public final class SimpleElasticJobTest extends EnabledJobIntegrateTest {
     
-    public OneOffDataflowElasticJobTest() {
-        super(new OneOffDataflowElasticJob());
+    public SimpleElasticJobTest() {
+        super(new FooSimpleElasticJob());
     }
     
     @Before
     @After
     public void reset() {
-        OneOffDataflowElasticJob.reset();
+        FooSimpleElasticJob.reset();
     }
     
     @Override
     protected JobConfiguration getJobConfiguration(final ElasticJob elasticJob, final String jobName) {
-        return JobConfiguration.newBuilder(jobName, JobType.DATAFLOW, 3).cron("0/1 * * * * ?").shardingItemParameters("0=A,1=B,2=C").misfire(false).overwrite(true)
-                .setProperty(DataflowJobExecutor.STREAM_PROCESS_KEY, Boolean.FALSE.toString()).build();
+        return JobConfiguration.newBuilder(jobName, JobType.SIMPLE, 3).cron("0/1 * * * * ?").shardingItemParameters("0=A,1=B,2=C").overwrite(true).build();
     }
     
     @Test
     public void assertJobInit() {
-        while (!OneOffDataflowElasticJob.isCompleted()) {
+        while (!FooSimpleElasticJob.isCompleted()) {
             BlockUtils.waitingShortTime();
         }
         assertTrue(getRegCenter().isExisted("/" + getJobName() + "/sharding"));
